Webapp Structure work
Restructured the webapp and created the build.xml file

+++ eclipse/common/src/com/digitald4/common/jdbc/HashCodeUtil.java
@@ -22,7 +22,7 @@ public class HashCodeUtil {
 	  * chars.
 	  */
 	  public static int hash( int aSeed, char aChar ) {
-	    return firstTerm( aSeed ) + (int)aChar;
+	    return firstTerm( aSeed ) + aChar;
 	  }
 
 	  /**
